Parse out facet-like options when building querystring for search
To make things easier on ourselves, anything that looks like a facet
is treated as a facet within the app (I'm looking at you range-limits!)
When the query string is constructed, however, facets with a `type`
property have their values added to the `options` object.

From what I can tell, the `blacklight_range_limit` is the only gem
we're using that will throw this wrench, so the fix is geared
primarily towards that, but should work with others.

Resolves #6.