round date range values to ensure even interval
`rc-slider` will throw a warning if `max - min` is not evenly
divisible by `step`. this rounds the `min` and `max` values to
the nearest `props.interval` when `FacetRangeLimitDate` is
first mounted as well as when `RangeSliderDate#onApplyRange` is
called. In the future, we may be able to remove it from the latter,
as (min + (interval * x)) and (max - (interval * x)) should always
be rounded (in theory).